chore: rename store/cart.js to guestCart
